CHORE : A&A filter 이름 수정
