+++ src/main/java/com/mju/insuranceCompany/global/jwt/JwtAuthenticationFilter.java
@@ -1,13 +1,16 @@
 package com.mju.insuranceCompany.global.jwt;
 
-import com.mju.insuranceCompany.service.user.service.UserService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mju.insuranceCompany.service.user.controller.dto.UserBasicRequest;
+import com.mju.insuranceCompany.service.user.domain.Users;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -15,33 +18,57 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@RequiredArgsConstructor @Slf4j
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
+@Slf4j
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+
+    private final AuthenticationManager authenticationManager;
     private final JwtProvider jwtProvider;
-    private final UserService userService;
+//    private final RefreshTokenService refreshTokenService;
+    private final AuthenticationFailureHandler authenticationFailureHandler;
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        // request에서 값을 가져온다
-        log.info("[JwtAuthenticationFilter] - filter 시작");
-        String authorization = request.getHeader("Authorization");
-        String token = null;
-        // token 변환
-        if (StringUtils.hasText(authorization) && authorization.contains("Bearer")) {
-            token = authorization.replace("Bearer","");
-            log.info("[JwtAuthenticationFilter] - 토큰변환 = {}",token);
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
+        log.info("[JWTAuthorizationFilter] - attemptAuthentication 시작");
+        ObjectMapper mapper = new ObjectMapper();
+        UserBasicRequest dto;
+        try {
+            dto = mapper.readValue(request.getInputStream(), UserBasicRequest.class);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
-        // 토큰 검증
-        if (token != null && jwtProvider.validateToken(token) != null) {
-            String loginId = jwtProvider.validateToken(token).getSubject();
-            log.info("[JwtAuthenticationFilter] - 토큰검증-로그인아이디 = {}",loginId);
-            SecurityContextHolder.getContext().setAuthentication(getAuthentication(loginId));
-        }
-        log.info("[JwtAuthenticationFilter] - filter 끝");
-        filterChain.doFilter(request,response);
+        UsernamePasswordAuthenticationToken authenticationToken
+                = new UsernamePasswordAuthenticationToken(dto.getUserId(), dto.getPassword());
+
+        return authenticationManager.authenticate(authenticationToken);
+
     }
-    private UsernamePasswordAuthenticationToken getAuthentication(String loginId) {
-        UserDetails users = userService.loadUserByUsername(loginId);
-        return new UsernamePasswordAuthenticationToken(users, users.getPassword(), users.getAuthorities());
+
+    // jwt 토큰을 만들어서 response로 전달해줘서 유저가 받아서 사용할 수 있게끔 한다.
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult) throws IOException, ServletException {
+        log.info("[JWTAuthorizationFilter] - successfulAuthentication 시작");
+        Users user = (Users) authResult.getPrincipal();
+        log.info("[JWTAuthorizationFilter] - User = {}", user);
+        String accessToken = jwtProvider.createAccessToken(user.getUserId());
+//        String refreshToken = jwtProvider.createRefreshToken(user.getUserId());
+
+//        TokenResponseDto dto = TokenResponseDto.builder()
+//                .accessToken(accessToken)
+//                .refreshToken(refreshToken)
+//                .build();
+//        refreshTokenService.createToken(user.getUsername(),refreshToken);
+
+
+        response.addHeader("Access-Token", accessToken);
+//        response.addHeader("Refresh-Token",dto.getRefreshToken());
+        response.setStatus(201);
+
+    }
+
+    @Override
+    protected void unsuccessfulAuthentication(HttpServletRequest request, HttpServletResponse response, AuthenticationException failed) throws IOException, ServletException {
+        log.info("[JWTAuthorizationFilter] - 인증실패");
+        authenticationFailureHandler.onAuthenticationFailure(request,response,failed);
     }
 }
